requirements.txt
# Core Dependencies
streamlit>=1.28.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0

# LangChain & LangGraph
langchain>=0.1.0
langgraph>=0.0.40
langchain-community>=0.0.20
langchain-openai>=0.0.8

# AI & ML
openai>=1.0.0
tiktoken>=0.5.0
sentence-transformers>=2.2.0
transformers>=4.35.0
torch>=2.0.0

# Vector Databases
chromadb>=0.4.15
faiss-cpu>=1.7.4
scikit-learn>=1.3.0

# PDF Processing
PyPDF2>=3.0.1
pdfplumber>=0.10.0

# Email & Communication
secure-smtplib>=0.1.1

# Visualization & UI
plotly>=5.15.0
plotly-express>=0.4.1
matplotlib>=3.7.0
seaborn>=0.12.0

# Data Processing
pydantic>=2.0.0
dataclasses-json>=0.6.0
typing-extensions>=4.7.0

# Utilities
requests>=2.31.0
python-multipart>=0.0.6
aiofiles>=23.0.0
asyncio>=3.4.3

# Development & Debugging (optional)
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Performance & Monitoring
psutil>=5.9.0
memory-profiler>=0.61.0

# Authentication & Security
cryptography>=41.0.0